api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
cryptography>=41.0.0
httpx==0.28.1
pydantic==2.10.4
tinydb==4.8.0
chromadb==0.5.23
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
beautifulsoup4==4.12.3
python-multipart==0.0.20
pynvml==11.5.0
psutil==6.1.1
pytest==8.3.5
pytest-asyncio==0.25.3
